+++ src/main/java/de/unistuttgart/iste/gits/flashcard_service/controller/FlashcardController.java
@@ -79,5 +79,4 @@ public class FlashcardController {
         return progressDataService.logFlashcardLearned(flashcardId, userId, successful);
     }
 
-
 }
