+++ src/main/java/de/unistuttgart/iste/gits/flashcard_service/controller/SubscriptionController.java
@@ -6,13 +6,8 @@ import io.dapr.Topic;
 import io.dapr.client.domain.CloudEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 import reactor.core.publisher.Mono;
-
-import java.util.Map;
 
 /**
  * REST Controller Class listening to a dapr Topic.
@@ -26,9 +21,16 @@ public class SubscriptionController {
 
     @Topic(name = "content-changes", pubsubName = "gits")
     @PostMapping(path = "/flashcard-service/content-changes-pubsub")
-    public Mono<Void> updateAssociation(@RequestBody CloudEvent<ContentChangeEvent> cloudEvent, @RequestHeader Map<String, String> headers) {
+    public Mono<Void> updateAssociation(@RequestBody CloudEvent<ContentChangeEvent> cloudEvent) {
 
-        return Mono.fromRunnable(() -> flashcardService.removeContentIds(cloudEvent.getData()));
+
+        return Mono.fromRunnable(() -> {
+            try {
+                flashcardService.deleteFlashcardSetIfContentIsDeleted(cloudEvent.getData());
+            } catch (Exception e) {
+                log.error("Error while processing content change event", e);
+            }
+        });
     }
 
 }
