+++ src/main/java/de/unistuttgart/iste/gits/flashcard_service/persistence/repository/FlashcardProgressDataRepository.java
@@ -1,6 +1,6 @@
 package de.unistuttgart.iste.gits.flashcard_service.persistence.repository;
 
-import de.unistuttgart.iste.gits.flashcard_service.persistence.dao.FlashcardProgressDataEntity;
+import de.unistuttgart.iste.gits.flashcard_service.persistence.entity.FlashcardProgressDataEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
