+++ src/main/java/de/unistuttgart/iste/gits/flashcard_service/persistence/repository/FlashcardRepository.java
@@ -1,9 +1,8 @@
 package de.unistuttgart.iste.gits.flashcard_service.persistence.repository;
 
-import de.unistuttgart.iste.gits.flashcard_service.persistence.dao.FlashcardEntity;
+import de.unistuttgart.iste.gits.flashcard_service.persistence.entity.FlashcardEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
@@ -15,7 +14,6 @@ import java.util.UUID;
  */
 @Repository
 public interface FlashcardRepository extends JpaRepository<FlashcardEntity, UUID>, JpaSpecificationExecutor<FlashcardEntity> {
-    @Query("select flashcard from Flashcard flashcard where flashcard.id in (:ids)")
     List<FlashcardEntity> findByIdIn(List<UUID> ids);
