+++ src/main/java/de/unistuttgart/iste/gits/flashcard_service/persistence/repository/FlashcardSetRepository.java
@@ -1,6 +1,6 @@
 package de.unistuttgart.iste.gits.flashcard_service.persistence.repository;
 
-import de.unistuttgart.iste.gits.flashcard_service.persistence.dao.FlashcardSetEntity;
+import de.unistuttgart.iste.gits.flashcard_service.persistence.entity.FlashcardSetEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
