+++ src/main/java/de/unistuttgart/iste/gits/flashcard_service/persistence/repository/FlashcardSideRepository.java
@@ -1,7 +1,7 @@
 package de.unistuttgart.iste.gits.flashcard_service.persistence.repository;
 
 
-import de.unistuttgart.iste.gits.flashcard_service.persistence.dao.FlashcardSideEntity;
+import de.unistuttgart.iste.gits.flashcard_service.persistence.entity.FlashcardSideEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
