+++ src/main/resources/graphql/common/ResourceMarkdown.graphqls
