+++ src/main/resources/graphql/service/flashcards.graphqls
