+++ src/test/java/de/unistuttgart/iste/gits/flashcard_service/test_utils/TestUtils.java
@@ -1,9 +1,7 @@
 package de.unistuttgart.iste.gits.flashcard_service.test_utils;
 
 import de.unistuttgart.iste.gits.common.resource_markdown.ResourceMarkdownEntity;
-import de.unistuttgart.iste.gits.flashcard_service.persistence.dao.FlashcardEntity;
-import de.unistuttgart.iste.gits.flashcard_service.persistence.dao.FlashcardSetEntity;
-import de.unistuttgart.iste.gits.flashcard_service.persistence.dao.FlashcardSideEntity;
+import de.unistuttgart.iste.gits.flashcard_service.persistence.entity.*;
 import de.unistuttgart.iste.gits.flashcard_service.persistence.repository.FlashcardSetRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Component;
